@@ -13,7 +13,9 @@
     "url": "http://brunoimbrizi.com"
   },
   "license": "MIT",
-  "sideEffects": false,
+  "sideEffects": [
+    "three"
+  ],
   "scripts": {
     "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
     "build": "webpack --config config/prod.js",
@@ -22,54 +24,60 @@
     "deploy": "gh-pages -d dist"
   },
   "dependencies": {
-    "@babel/preset-typescript": "^7.12.7",
     "async-preloader": "^4.9.2",
     "cannon-es": "^0.15.1",
+    "core-js": "^3.8.2",
     "domready": "^1.0.8",
-    "postprocessing": "^6.17.4",
+    "postprocessing": "^6.19.0",
     "rxjs": "^6.6.3",
     "stats.js": "^0.17.0",
-    "three": "^0.123.0",
-    "tsconfig-paths-webpack-plugin": "^3.3.0",
+    "three": "^0.124.0",
     "tweakpane": "^1.5.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.6",
+    "@babel/core": "^7.12.10",
+    "@babel/plugin-proposal-class-properties": "^7.12.1",
+    "@babel/plugin-proposal-object-rest-spread": "^7.12.1",
     "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@babel/polyfill": "^7.11.5",
-    "@babel/preset-env": "^7.11.5",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-typescript": "^7.12.7",
     "@types/domready": "^1.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.12.0",
+    "@typescript-eslint/parser": "^4.12.0",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.1.0",
+    "babel-plugin-syntax-dynamic-import": "^6.18.0",
+    "babel-plugin-transform-imports": "^2.0.0",
     "clean-webpack-plugin": "^3.0.0",
-    "copy-webpack-plugin": "^6.2.1",
+    "copy-webpack-plugin": "^7.0.0",
     "css-loader": "^5.0.1",
-    "eslint-config-airbnb": "^18.0.1",
-    "eslint-config-prettier": "^7.0.0",
-    "eslint-config-react": "^1.1.7",
-    "eslint-plugin-prettier": "^3.1.2",
+    "eslint": "^7.17.0",
+    "eslint-config-babel": "^9.0.0",
+    "eslint-config-prettier": "^7.1.0",
+    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-webpack-plugin": "^2.4.1",
     "extract-loader": "^5.1.0",
     "file-loader": "^6.1.1",
     "gh-pages": "^3.1.0",
     "glslify": "^7.1.1",
     "glslify-import-loader": "^0.1.2",
     "glslify-loader": "^2.0.0",
-    "html-webpack-plugin": "^4.5.0",
+    "html-webpack-plugin": "^4.5.1",
     "imports-loader": "^1.2.0",
     "postcss-loader": "^4.0.4",
     "postcss-preset-env": "^6.7.0",
     "prettier": "^2.2.1",
     "raw-loader": "^4.0.2",
-    "sass": "^1.27.0",
+    "sass": "^1.32.0",
     "sass-loader": "^10.0.3",
     "style-loader": "^2.0.0",
-    "terser-webpack-plugin": "^5.0.3",
-    "ts-loader": "^8.0.12",
-    "typescript": "^4.1.2",
-    "webpack": "^5.10.0",
+    "ts-loader": "^8.0.14",
+    "tsconfig-paths-webpack-plugin": "^3.3.0",
+    "typescript": "^4.1.3",
+    "webpack": "^5.11.1",
     "webpack-bundle-analyzer": "^4.3.0",
-    "webpack-cli": "^4.2.0",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-merge": "^5.2.0"
+    "webpack-cli": "^4.3.1",
+    "webpack-dev-server": "^3.11.1",
+    "webpack-merge": "^5.7.3"
   }
 }
