@@ -4,16 +4,14 @@ const { CleanWebpackPlugin } = require("clean-webpack-plugin")
 const CopyWebpackPlugin = require("copy-webpack-plugin")
 const HtmlWebpackPlugin = require("html-webpack-plugin")
 const TsconfigPathsPlugin = require('tsconfig-paths-webpack-plugin');
-
 const __root = path.resolve(__dirname, "../")
 
 module.exports = {
   entry: {
     index: [
-      "@babel/polyfill",
       "./src/scripts/index.ts",
       "./src/styles/main.scss",
-    ],
+    ]
   },
   output: {
     path: path.resolve(__root, "dist"),
@@ -23,21 +21,15 @@ module.exports = {
   module: {
     rules: [
       {
-        test: /\.tsx?$/,
-        use: "ts-loader",
-        exclude: /node_modules/,
+        test: /\.jsx?$/,
+        use: ['babel-loader'],
+        include: path.resolve(__root, 'src', 'scripts'),
       },
-      /*{
-        test: /\.js$/,
-        use: {
-          loader: "babel-loader",
-          options: {
-            presets: ["@babel/preset-env"],
-            plugins: ["@babel/plugin-syntax-dynamic-import"],
-          },
-        },
-        exclude: /node_modules\/(?!(postprocessing)\/).*!/,
-      },*/
+      {
+        test: /\.tsx?$/,
+        use: ['babel-loader', 'ts-loader'],
+        include: path.resolve(__root, 'src', 'scripts'),
+      },
       {
         test: /\.(glsl|frag|vert)$/,
         use: [
@@ -83,11 +75,12 @@ module.exports = {
   resolve: {
     extensions: [".tsx", ".ts", ".js", ".scss"],
     alias: {
-      "@core": path.resolve(__root, 'src/scripts/_core/'),
+      "@src": path.resolve(__root, "src/"),
+      "@core": path.resolve(__root, "src/scripts/_core/")
     },
     plugins: [
-      new TsconfigPathsPlugin()
-    ]
+      new TsconfigPathsPlugin(),
+    ],
   },
   plugins: [
     new CleanWebpackPlugin(),
